requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.5
blinker==1.4
cffi==1.11.5
Click==7.0
cryptography==2.4.2
enum34==1.1.6
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.1.5
Flask-WTF==0.14.2
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
passlib==1.7.1
pycparser==2.19
PyMySQL==0.9.3
pytz==2018.7
six==1.12.0
speaklater==1.3
SQLAlchemy==1.2.15
Werkzeug==0.14.1
WTForms==2.2.1
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.5
blinker==1.4
cffi==1.11.5
Click==7.0
cryptography==2.4.2
enum34==1.1.6
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.1.5
Flask-WTF==0.14.2
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
passlib==1.7.1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
speaklater==1.3
SQLAlchemy==1.2.15
sqlalchemy-datatables==1.2.0
Werkzeug==0.14.1
WTForms==2.2.1
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.5
blinker==1.4
cffi==1.11.5
Click==7.0
cryptography==2.4.2
enum34==1.1.6
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.1.5
Flask-WTF==0.14.2
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.17.0
marshmallow-sqlalchemy==0.15.0
passlib==1.7.1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
speaklater==1.3
SQLAlchemy==1.2.15
sqlalchemy-datatables==1.2.0
Werkzeug==0.14.1
WTForms==2.2.1
